Fix album browser's border (bgo#644315)
On Windows, the opacity of the second line of text (the artist) being set
to != 1.0 was causing the border to not get drawn around the entire album grid.

This is caused by the cr.Push/PopGroup calls, though I'm not sure why; likely
just an issue with our currently bundled cairo binary.