libbanshee: Correctly detect sink volume control (bgo#644648)
If an element is used as an audiosink which detects and dynamically
plugs the actual sink element, then we can only detect sink volume
support after this has happened. To trigger this, simply set the sink
state to READY, which should open the resources and thus load the actual
audio sink.

Signed-off-by: Gabriel Burt <gabriel.burt@gmail.com>