PlayerEngineService: Obtain library or metadata info for opened uris
When opening a file from a uri, instead of creating an unknown track,
try to find the uri on the library or extract the metadata from the
file.
This only affects the Open method used by the D-Bus interfaces. The
various other ways to open a file use different code paths.

Fixes bgo#636071.

Signed-off-by: Bertrand Lorentz <bertrand.lorentz@gmail.com>