Banshee.Services: Avoid creating dupes after moves (bgo#638889)
When importing, besides checking that there is not already
an entry with the same Uri, now we check also for the same
MetadataHash. If there is an entry which points to an URI
that doesn't exist anymore and have the same hash for a file
to be imported, we consider it outdated, and we update its
URI with the one of the file that was going to be imported,
so we avoid a broken duplicate and we keep the same
DateAdded value that we had originally.