Update the release process for the tarball installation on ftp.gnome.org
Tarballs are uploaded to $GNOME_FTP_USER's home folder on
master.gnome.org. You have to manually ssh to the server and then run
"ftpadmin install ....tar.bz2" to make the tarball available on
ftp.gnome.org. Maybe this could be automated, but I'm not sure it's a
good idea.

In fact we only need to upload just one tarball, ftpadmin is supposed to
take care of the rest: generated tarballs with the other compression
type (.gz or .bz2), publish NEWS, generated checksums, etc.

So make-release can probably still be improved and streamlined.