SourceManager: Add and delete group sources as necessary (bgo#639805)
Whenever a source is added or removed, SourceManager checks its list of
GroupSources to determine if a GroupSource needs to be added or removed
from the source list. To achieve this, the SourceManager keeps track of
GroupSources in a separate list in addition to the main sources list.

DapService is also modified to reflect this generalized approach. The
GroupSource is only added once when the DapService is initialized. From
there, the SourceManager decides whether or not to display the
GroupSource depending on whats added/removed.

Signed-off-by: Bertrand Lorentz <bertrand.lorentz@gmail.com>