Mpris: Use PlayerEngine to get the current track (bgo#648498)
As PlaybackController.CurrentTrack is updated when the StartOfStream is
raised, there is no guarantee that it will be up-to-date when the Mpris
extension processes that same event. PlayerEngine.CurrentTrack is
updated before the event is raised, so it will always contain the track
that is starting to play.

This will prevent the Mpris extension from sending the metadata from the
previous track with the PropertiesChanged signal.