SmartPlaylists: Let child sources decide if the parent can add them
SmartPlaylists, when loaded from the database, are added to their
parent source after they have been initialized, refreshed and
reloaded. The consequence of this is that it was completely up
to their parent if it was OK to add them as children of not.
However SmartPlaylistSources can have an setting enabled called
HiddenWhenEmpty, which is what the default/predefined smart
playlists have activated to prevent polluting the UI (especially
for new users that haven't imported anything yet), so this was
not taken in account and causing BGO#657479.

The alternative fix to this would be to just call Reload() on the
playlist after it has been added to the parent, but that could cause
extra overhead and this is a more elegant solution: returning bool for
SetParentSource so children can veto their addition to the parent.

We also get the side-benefit of having the "virtual" keyword on
this method be actually useful, because before this commit,
SmartPlaylistSource was the only derived class that was overriding
it, but calling just the base method, so indeed not overriding it
but having redundant and unnecessary code.

Signed-off-by: Bertrand Lorentz <bertrand.lorentz@gmail.com>