Bookmark: Fix loading of bookmarks so that playing starts
If we only call PlayerEngine.Open() we stay at the Loading state and
PlayerEngine.CanSeek never becomes true, so we can't seek to the
bookmark location.

Calling PlayerEngine.OpenPlay() makes sure we reach the state where we
can seek and the track actually plays. This might bring back a bit of
stutter, if we happen to start playing before seeking, but it's better
than not playing at all.