Banshee.Services: Refactoring, BansheeQuery.GetSort to receive QueryField
BansheeQuery.GetSort() method now receives a QueryField object
instead of a string with the key to be sorted. This way we
stop hardcoding some columns that were already defined in the
corresponding QueryField instances, and we replace the big
string-swtich statement with object-reference comparisons,
which is faster, although still not perfect (I guess this can
still be improved with a bit of polymorphism, maybe moving
the method to QueryField class, and overriding in the special
cases).

This commit also removes a TODO and a FIXME. Win.