Metadata: Split WriteRatingsAndPlayCount into two options
The configuration for embedding Ratings and Playcounts
into files was not separated before. In certain
situation this is not desirable because you may want
to make the files change to propagate the Rating to
other computers, in case you share your media collection
somehow, but propagating the tracks everytime they
play is too expensive. Fixes bgo#602159.