Add initial support for DVD playing
Add the necessary bits to support DVD playing in both libbanshee and
GStreamerSharp, including menu navigation.

Rework the AudioCD extension into an OpticalDisc extension that supports
both audio CDs and video DVDs.

Adapt the NowPlaying extension to forward mouse and keyboard events to
the player engine when appropriate, to implement menu navigation for
DVDs.

This is the result of a collaboration between Alex Launi and Olivier
Dufour, with a few changes and cleanups by me.