Muinshee: Make the Esc key close the dialog if there is no search text
Override the KeyPress handler of the search entry to have the Esc key
press bubble up if the search entry is empty. In this case the dialog
will be closed. Fixes bgo#611333.

The behavior now matches common search entries: if the entry contains
text, the first hit on Esc clear the text, the second closes the dialog.

Signed-off-by: Bertrand Lorentz <bertrand.lorentz@gmail.com>