BansheeQuery: stop hardcoding field columns
After some recent refactorings to start using QueryField
objects instead of strings, it makes sense to improve
further and stop using in general column names that are
already specified in these field objects. Leads to cleaner
code, less duplication, and less error-prone to typos as
it is checked at compile time now.

There is no change of behaviour in this commit.