Gio: Fix encoding of URIs in Directory class
The Gio IO backend was buggy, since we started to import filenames with
some special characters like ',' and '&' escaped.

So this would lead to URIs like this:

file:///path/You've%20Come%20a%20Long%20Way,%20Baby/Right%20Here%2C%20Right%20Now.mp3

Which is inconsistent because:
- Commas are encoded in the file name while the normal IO banshee
backend never did this.
- Commas are encoded in the file name but not in the directory name.

In fact, SafeUri's default behaviour when asked for the AbsoluteUri
property is encoding some characters, but not so many as
System.Uri.EscapeDataString() method appears to do.

So with this fix in place, the URI will be now:

file:///path/You've%20Come%20a%20Long%20Way,%20Baby/Right%20Here,%20Right%20Now.mp3

This failure to encode URIs consistently in the database was causing
user-facing issues like bgo#631970 and bgo#661100, which will be
completely fixed after the next commit, that fixes the existing URIs in
the database.

Signed-off-by: Bertrand Lorentz <bertrand.lorentz@gmail.com>