Dap.MassStorage: don't sync 'unknown album' metadata
When a sync via MassStorage Dap happens,
tracks without album info or singles were
wrapped in a subfolder called "Unknown Album"
in the language that Banshee is translated to,
or a folder with this term in the name.

This is confusing because "Unknown Album"
is just a placeholder used by Banshee to
present data in a more aesthetic way when
it is missing, to also let the user find
it more easily. The OS of the device should
instead be the one which decides how to
handle this data being missing. In this
case we only adapt MaemoDevice for this
change because it has a FolderDepth of 2,
but can also work with a FolderDepth of 1.

For devices with FolderDepth < 2, this
will be fixed automatically. For devices
with FolderDepth == 2, their corresponding
*Device.cs class should be modified in
case it works properly.