configure: clearer check for AM_GCONF_SOURCE_2 macro
The AM_GCONF_SOURCE_2 macro was used in the configure
phase without adding a previous check for the package
that contains it (called 'gconf2-devel' in most distros).

This was causing either autoconf complaining for not
finding it (which is solved by using m4_pattern_allow)
before a PKG_CHECK_MODULES() can be run when autogen.sh
was called and the 'configure' script was not generated
yet, or by the configure script itself if the call to
PKG_CHECK_MODULES(LIBGCONF, gconf-2.0) was not done
before the call to the macro.

This change comes with the extra of wrapping all gconf
configure-checking logic into a banshee m4 macro like
it is done for other dependencies.

Fixes bgo#604416