DatabaseAlbumArtistListModel: Fix type-ahead find with album artists
DatabaseAlbumArtistInfo uses the CoreArtists table, but
DatabaseAlbumArtistListModel was using a query field based on the
CoreAlbums table, which broke the query used for type-ahead find.

Use the ArtistField query field instead, which is based on the
CoreArtists table, and adapt the reload sql fragment to start from
CoreArtists. Fixes bgo#664933.