Fix importing of rating and play count when updating track
For tracks that were already imported, the rating and play count would
never get updated from the file when rescanning the library or updating
through the library watcher.

We now look at those two tags during those operations, but only if the
corresponding "Write ... to file" is set. If the option is not set, it
means the value in the database is probably more up-to-date than the one
in the file, so we don't import it from the file, to avoid data loss.

Fixes bgo#638130 and bgo#665963.