Audiobook: Add a button to go back to the list of audiobooks
Previously, the button to go back to the list of audiobooks (aka
GridView) was in the TitleWidget. That widget is not displayed anymore
in the UI, so we can remove the code using it.

Add a "Go to Audiobooks" button in the SourceActions placeholder, to
replace that lost UI and allow users to go back to the GridView. Fixes
bgo#656439.