DatabaseTrackInfo: reuse another SQL query
Reuse another SQL instruction for querying
the TrackID filtering by Uri. The UriField is
now also used, which prevents hardcoding the
Uri column name.

This also prevents that someone could misuse
the GetTrackIdForUri method by passing null
or an empty array, as the primary_sources param
was not validated before.

Signed-off-by: Alexander Kojevnikov <alexk@gnome.org>