DapSource: Flush pending changes when disposing too (bgo#666696)
The code to flush pending sync changes was only being called when the
user chose to eject the device. However this should not be the only
operation that yields the flush, as quitting the media player is not
meant to discard automatically the changes that a user may have done to
a Dap source.

Ejecting a device causes the corresponding DapSource to be disposed
later, so we need to make sure we don't do the flush twice, especially
as by the time DapSource is disposed after ejecting, the volume is
already unmounted.

Signed-off-by: Bertrand Lorentz <bertrand.lorentz@gmail.com>