Fix race condition when starting multiple instances (bgo#648522)
We were first only checking whether the DBus name had an owner, and
grabbing the name much later in the startup process. This created a race
condition, so two Banshee processes started at almost the same time
could both think they were the primary instance.

Add a DBusConnection.GrabDefaultName method which tries to grab the
default DBus name and returns whether we are the primary instance or
not. If DBus is not available we have to assume we're the primary
instance. Use that new method in Booter.