Dap.MassStorage: metadata resync in manual mode (bgo#589196)
Add support for changing the metadata of the tracks in
MassStorage devices in manual mode, making sure the path
of the file is updated if needed and that the changes to
the file get written thanks to SaveTrackMetadataJob.

Playlists will get re-written with the changes (if any)
when the user ejects the device (as explained in the
comment "Track transfers happen immediately, but
playlists saves don't" in DapSource class, Eject method).
This is to avoid that the sync operation rewrites each
playlist once per track (in case of once per playlist).