Fix MusicBrainz dllmap for cross-architecture support
Not every platform Banshee runs on uses libc.so.6 - for example, Itanium uses
libc.so.1, and kFreeBSD (on Debian) uses libc.so.0.1.

This is actually not a big deal, because /etc/mono/config contains an explicit
Dllmap for this case, which is populated with the host platform's correct
soname at build time. For example, on Itanium:

	<dllmap dll="libc" target="libc.so.6.1" os="!windows"/>

and on kFreeBSD-AMD64:

	<dllmap dll="libc" target="libc.so.0.1" os="!windows"/>

You'll notice, however, that this mapping only works when the target is
"libc" - and the MusicBrainz library in Banshee targets "libc.so.6"

Changing this to "libc" fixes execution on exotic architectures without
hurting existing architectures, thanks to the mapping in /etc/mono/config

Signed-off-by: Alexander Kojevnikov <alexk@gnome.org>