UPnPService: Handle UPnP device removal
When a UPnP device goes away, the corresponding sources are now removed.
To do this, we keep track of the association between each device UDN
(which is a UUID) and the corresponding UPnPServerSource.

The "Shared Media" container source is now only displayed if there
something under it. This is the same behavior as the DAAP container
source.