WebBrowser: Implement SSL certificate validation (bgo#671536)
Add a SecurityLevel property to OssiferWebView that indicates the result
of the SSL certificate validation.

For now, any page loaded using HTTPS with an invalid or untrusted
certificate will be blocked, and an ugly message will be displayed
instead.

This commit breaks our string freeze by adding two strings, but it
couldn't be avoided.