windows: Don't show the EULA dialog in the installer (bgo#647210)
We do not need the user to agree to the MIT license or the GPL, so we
don't need to show them. Saving your mouse, one click at a time...

To achieve that we need to replace the standard WixUI_InstallDir UI
module with its definition, and adapt it to remove the license agreement
dialog.

We now install the license.rtf file to the root of the install folder,
which fixes bgo#642818. The first paragraph of the file is also adapted
for clarity.