AppleDeviceTrackInfo: Fix build against libgpod-sharp 0.8.2
In libgpod 0.8.2, released in July 2011, the type of the Track.Size
field was changed, so it is reflected in the C# bindings and breaks the
API.

We have currently no way of checking which version of the libgpod-sharp
bindings are available, so this breaks compilation against older version
of libgpod-sharp. But people shipping Banshee 2.5.x and later should
really ship libgpod-sharp 0.8.2.

See this bug for details and attempts to fix this correctly:
https://bugzilla.gnome.org/show_bug.cgi?id=655660