MassStorage: Save playlists relative to BaseDirectory (bgo#587756)
Until now playlists were created under the first folder in AudioFolders
or BaseDirectory if AudioFolders wasn't set.

The (now obsolete) HAL specification described playlist_path as being
relative to the mount point of the device.

The Media player info repository doesn't explicitly describe what is
the base directory for playlist_path. But common values show that it's
relative to the root of the device.

So we now consider the PlaylistPath to be relative to the base directory
of the device. Default values for the devices with custom support are
adjusted to also reflect that.

Signed-off-by: Bertrand Lorentz <bertrand.lorentz@gmail.com>