MassStorage: Support reading playlists from multiple folders
The PlaylistPath property in media-player-info files can in fact contain
a list of folders. So we now parse it as a list, and look for playlists
in all those folders.

Playlists are written only to the first folder in that list, similar to
what we do with AudioFolders. As we systematically regenerate all
playlists, that means we might move playlists from other folders to that
first folder. That should probably be improved in the future.

The playlist_path property for our .is_audio_player file still contains
a single value. Same thing for custom device support, which only has to
provide a single DefaultPlaylistPath.