osx: Switch from ige-mac-integration to gtk-mac-integration (bgo#674576)
Update DllImport references to new gtkmacintegration libs and update the
ige_* symbols to gtk_* symbols.

We're still using the old (deprecated) Carbon-based gtk-mac-integration
API instead of the proposed GtkOSXApplication API, which resides in the
same library.

Also update Hyena to the latest git master, to get rid of unneeded
references to libigemacintegration.

Signed-off-by: Bertrand Lorentz <bertrand.lorentz@gmail.com>