Migo.Syndication: Handle feeds with malformed enclosure URLs
When parsing an RSS feed, don't create an enclosure if the URL is empty.
We were onyl considering when the URL is missing, but a present but
empty value is just as useless.

Catch exceptions when downloading an enclosure. This prevents us from
crashing when we already have an empty enclosure URL in the database.

Public announcement: Do yourselves a favor and go and read one of Cory
Doctorow's books, available at http://craphound.com/