FileSystemQueueSource: Reload when tracks are imported (bgo#634063)
When you drag'n'drop tracks from the FSQ to the music library, they are
imported and are not in the FSQ anymore. But the FSQ was not refreshed,
so they still appeared there, and clearing the FSQ would delete them
from the library.

Make sure the FSQ is refreshed when tracks are imported, so that the
content of the FSQ is always up-to-date.