ServiceManager: check properly if a service is required (bgo#626659)
There was infrastructure already in Banshee to determine if certain
service initialized could be ignored or not: services which are
fundamental to properly run Banshee where just marked with the
interface IRequiredService. However the check for this was wrong,
as it tried to check the type of the service instance, rather than
the service type; then, the instance would be null if it couldn't
be initialized, and a null value apparently cannot behave as we
want when used with the 'is' operator.

The check now is done against the type of the service that needs
to be initialized, rather than the instance resultant of the
initialization.

This bug was causing fatal errors to be ignored and allowing the
resuming of Banshee launch, which in the end caused other kind
of exceptions unrelated to the root cause (i.e. NREs when the
database file is locked or gets corrupted).

Signed-off-by: Bertrand Lorentz <bertrand.lorentz@gmail.com>