build: Don't run update-desktop-database in an empty folder
During "make distcheck", update-desktop-database creates a cache file
which we have no reasonable way to remove. This left-over file now
breaks distcheck with recent automake versions.

We now only run update-desktop-database if DESTDIR is not set and if
there's already default mime-types in the target folder. That means it
will not be run when doing "make install" into an empty folder.