PrimarySource: Fix pruning of artists from the database (bgo#677288)
The existing query to prune artists from the CoreArtists table was
only keeping artists referenced in CoreTracks, thus removing all album
artists that were not also track artists.

Fix this query to also keep artists referenced in CoreAlbums. Also
refactor the query to use NOT EXISTS (...) instead of NOT IN (...),
which is much faster with big databases.

In addition, prune the CoreAlbums table first, so that deleted album
artists will be pruned from CoreArtists by the second query.

For some reason, for the CoreAlbums table, the NOT EXISTS (...) query
has slightly worse performance, so we keep the NOT IN (...) query.
Apparently, query optimization is more of an art than a science...