LastFM: Clean the queue after a scrobbling failure (bgo#667499)
If invalid tracks are present in the audioscrobbler queue we receive
an error when submitting to LastFM. Since we are not cleaning the
queue upon a failure we continue to attempt to resubmit indefinitely.

Adds a step to remove invalid tracks from the queue upon receiving
a failure so this no longer occurs. Also adds a simple logging
statement when invalid data is detected in a track being added to
the LastFM queue to help diagnose the root cause of the submission
failures.

Signed-off-by: Alexander Kojevnikov <alexk@gnome.org>