OSX: Allow more actions to have custom accelerators (bgo#677866)
If an action is initialized with null as an accelerator, it can't be
customized. So we initialize the accelerator to "" for
PreferencesAction, TrackPropertiesAction and NewSmartPlaylistAction, so
that we can use the custom OSX accelerator map for those actions too.

This means that the stock accelerators for those actions will not be
used anymore, but apparently they don't have any anyway.

Signed-off-by: Bertrand Lorentz <bertrand.lorentz@gmail.com>