OSX: move to cocoa API and enable keybindings
This commit moves from the ige-/gtk-mac-integration API to the
GtkOSXApplication API (which is still provided by the same library
called 'gtk-mac-integration', not to be confused with the API name).
This means banshee no longer uses any OS X carbon functions (which are
deprecated), but cocoa only. The Banshee.Osx backend also now fully
relies on gtk-mac-integration as interface to OS X specific functions
and does no longer contain own DllImports into OS X core framework.

This drops the support for closing the banshee main window (via
Cmd+Close) since GtkOSXApplication does not handle this properly right
now. Banshee is a single-window only application at the moment, so this
shouldn't be a usecase anyway.

Additionally, this commit adds OS X specific keyboard accelerators which
are modeled after iTunes. Upon first run on OS X, a custom osx_accel_map
is copied to the configuration dir, that is loaded on startup and
overwrites the banshee default accelerator map. To allow binding to some
actions which are not bound by the default accelerator map, i've
modified those to init with an empty mapping ("" instead of null).