Banshee.IO: Use strongly-typed SafeUri for GetFiles and GetDirectories
Fix a FIXME that was placed in the Banshee.IO.IDirectory interface:
the files and directories returned by all backends should be returning
the same format of filesystem representation. By using SafeUri type,
we enforce this check at compile time.

Thanks to this refactoring, bugs like bgo#679252 are fixed and should
not happen again.

Signed-off-by: Bertrand Lorentz <bertrand.lorentz@gmail.com>