osx: fix muted volume on startup (bgo#660012)
This commit fixes a bug in the libbanshee library, which did not set the
initial current_volume. As a result, the volume returned by
bp_get_volume () will always be 0 until one modifies/sets a new volume
via the volume slider.

This bug also manifests itself on Linux, but the
volume would slider get set to the right value once you start playing a
track.

Signed-off-by: Bertrand Lorentz <bertrand.lorentz@gmail.com>