Replace custom volume widget with gtk's VolumeButton (bgo#681062)
We implemented a custom volume slider gtk widget which was not present
at that time inside gtk. However, since 2.12, gtk ships a volume widget,
which is also bound by gtk-sharp as Gtk.VolumeButton.

This patch modifies ConnectedVolumeButton to derive from
Gtk.VolumeButton, thus the custom widget code can be removed. It also
removes the "Classic" field of ConnectedVolumeButton as it was used to
display the scale above the button so that it didn't go off-screen, and
this is now done automatically by VolumeButton.

ConnectedVolumeButton now has a Disconnect method, used by
FullscreenControls, because VolumeButton doesn't like to have its value
changed after it has been destroyed.

Signed-off-by: Bertrand Lorentz <bertrand.lorentz@gmail.com>