SourceView: replace a dynamic cast with a static cast
While looking at the source code to find potential reasons for
bgo#683359 I found this usage of "as" keyword that did not have
a subsequent check for null. This could cause NullReferenceExceptions
instead of the more meaningful InvalidCastExceptions.

(Although this change is in line 141, not line 137 like the stack
trace in the bug points out.)