SourceView: log errors in SetCellDataFunc to avoid crashing (bgo#683359)
There must be a tricky problem about model.GetValue() not returning
proper values the first times it is queried. It is an issue hard
to reproduce so this is not a proper fix, but just rather a call
to Log.Error to prevent crashing.

(Banshee is able to run normally after a few times this error
condition happens and gets logged, as tested by Andy Goar.)

Signed-off-by: Bertrand Lorentz <bertrand.lorentz@gmail.com>