PathPattern: Limit filename length to 200 characters (bgo#636450)
If a track has a really long title, we would reach the limit of the
filesystem and fail to write it. In particular, this lead to crashes
when ripping some CDs.

On most filesystems, filenames are limited to 255 characters or 255
bytes. As we can't be sure of the byte size when the filename reaches
the filesystem, we use an arbitrary limit of 200 characters. This should
be enough to be safe in most cases, without being too limiting.

Signed-off-by: Bertrand Lorentz <bertrand.lorentz@gmail.com>