AppleDeviceSource: refactoring, extract submethods from PerformSyncThreadCycle
The method PerformSyncThreadCycle() was more than a hundred lines long and it
was easy to get lost on it. Now reduced to 33, make the code much more readable,
and stacktrace-friendly in case of exceptions pasted in bug reports.

No change of behaviour in this commit.

NB: Had to choose the too long name SyncTracksToPlaylists because there is already
a SyncPlaylists method in the class! (higher level than this one).