AppleDeviceSource: postpone the removal of invalid tracks to the sync cycle
The fix for bgo#634652 has caused a bit of trouble for other users:
a) It caused bgo#679260, which was fixed in Banshee 2.5.0.
b) It delayed the initialization of the device, as explained by a user
in https://bugzilla.gnome.org/show_bug.cgi?id=679260#c25.

To avoid b), we now postpone the removal of invalid tracks to the sync cycle.

This way:
- There is no delay in the load process.
- No writing operations happen in the device until there is a need of a sync.

In this commit we also wrap the deletion under a try-catch, to be super sure that
the sync process can continue (in the same way it is done for the removal of non
invalid tracks).