AppleDeviceSource: Improve previous commit, to dequeue, not enumerate
A foreach loop doesn't need a pre-Count check really. But, wait,
invalid_tracks_in_device is now a Queue like the other collections
to sync, so the proper way to do this is Dequeuing, not enumerating.

...reviewer from hell said.

NB: lock is not needed here when dequeueing because invalid tracks
cannot be added by the user on the fly while a sync happens.