PlaylistElement: make this class strongly-typed
This class will no longer be Dictionary-based. Before this, it was a
Dictionary<string,object> in order to have essentially a list of elements
with different types, accessible through strings. Now these elements are
just the 3 that were being used (Uri, Duration, Title) but strongly-typed
(to Uri, TimeSpan and string respectively).

This avoids the use of "object" as effectively a dynamic type, and thus
allows us to remove some nasty casts that were in place, increasing type
safety, performance (micro?)optimization and readability.

No change of behaviour in this commit.