YouTube: advice about --disable-youtube configuration flag (bgo#686582)
YouTube extension is an optional requirement to build banshee, however
it was not very clear how to disable it if the dependencies required
for it were not found. With this modification we get also a bit more
consistent to how other extensions advice about their --disable flags
(i.e.: AppleDevice).

This has the additional benefit of simplifying a bit youtube's m4 macro,
as there is no longer a nested PKG_CHECK_MODULES call inside another one.