build: Make gio-hardware be enabled by default (bgo#686584)
Before this change, at config time this M4 would determine if the GIO
Hardware backend (now the default and unique hardware backend) would be
built depending on the package dependencies present in the system.

Problem is, if dependencies for it were not met, the default was to
silently avoid building the backend.

This commit changes this so it is now mandatory to pass the flag
--disable-gio-hardware if the user wishes to compile without it.

It doesn't completely solve bgo#686584 but helps alleviate the problem.
And has the added benefit of making more versions of dependencies be
clearly visible at the top of the macro.