configure: better guess for the C# compiler available
There are some distributions that bundle only a limited set of class
libraries to be able to meet size requirements better. Such is the case
of Ubuntu, which normally ships with only the subset compiled for the
4.0 version of the runtime.

In this case, the user may install the default Mono SDK packages when
asked by the configure script, which only includes the 4.0 compiler by
default (dmcs). At that point, banshee can already be compiled, but
the configure script would insist in requiring a 2.0 compiler. If the
user installs it, compiles Banshee and launches the "run" target, she
may get a runtime exception because of a runtime version mismatch, like
in bgo#675113.

This problem is commonly known, and there is a clear workaround for it:
http://orangesquash.org.uk/~laney/blog/posts/2011/10/mono-gotcha/
However it poses a small barrier of entry for new developers that try
to compile and run banshee from sources for the first time.

Now the configure script is smarter and tries to guess harder if there
is a C# compiler, even one that targets a higher version of the runtime
than the one that is required to compile banshee.

This possibility is only enabled in case the Mono version found is 2.8
or higher, because Mono 2.6.1 included the 4.0 profile only as a
"preview" [1] and might not be stable enough.

[1] http://www.mono-project.com/Release_Notes_Mono_2.6.4#C.23_Language