build: use the gui-thread-check profiler module by default when debugging
Upcoming versions of gtk# (2.99.1 and higher) will bundle this useful
profiler module by default. To reduce the feedback loop between developers
and bug reporters, banshee will run it whenever launched with the --debug
flag or when running in uninstalled mode (via "make run").

This way the logs that we get will include any traces of GUI calls that
were incorrectly not sent to the main thread, which in occassions can cause
crashes or unexpected behaviour.

Even though this is not bundled in gtk# 2.x series, it may still useful to
try to run it at launch, in case the user has installed the profiler module
on his own, or advised by devs: https://github.com/slluis/gui-thread-check

Signed-off-by: Alexander Kojevnikov <alexander@kojevnikov.com>