RescanPipeline: receive a Func<> to warn about high removal (bgo#577225)
Even though the current Rescan operation for the music library doesn't
remove (for now) non-existent tracks that are not inside the default folder
of its source (normally $HOME/Music ), there could be a possibility that
a user set up some symlinks inside this folder that linked to a removable
drive, or that the mount point was inside this subtree.

In those cases, whenever a Rescan operation was done (either by the menu
or triggered by the LibraryWatcher extension), all tracks would be removed
from the library, which would be especially bad if there was data that the
user couldn't recover (for example if the user kept its ratings outside the
files), because we don't have yet any automatic DB-backup mechanism.

This change fixes this scenario by prompting the user a confirmation dialog
that would pop up if the amount of tracks to be removed would be greater
than 50.

Signed-off-by: Andres G. Aragoneses <knocte@gmail.com>