build: test dmcs before gmcs when trying to find the compiler (bnc#798267)
In the previous improvement that had been done [1] in this mono.m4 macro
to find the 4.0 profile compiler in case it's mature enough (mono 2.8),
the scenario in which both compilers could be installed was not considered.

Now in this case, a higher version of the compiler (dmcs) will be used.

[1] http://git.gnome.org/browse/banshee/commit/?id=3d3f7899465ee46e634b53c236401f612dac1492

Fixes BNC#798267 (https://bugzilla.novell.com/show_bug.cgi?id=798267)