Dap.MassStorage: not import dirs twice for case-insensitive drives (bgo#653282)
Case insensitive file-systems, when imported via the MassStorage extension,
could include duplicate entries because the default folders configured could
contain directory names that were equivalent, i.e. /Music and /music.

A more thorough check could have been done creating a file and checking that
the file content is the same even if queried with a different case, but:
- That would require read/write access, which we cannot guarantee.
- Is less safe (could overwrite an existing file) and less elegant.

A list of case insensitive file-systems types that Banshee is aware of, is
good enough approach then.

Signed-off-by: Andres G. Aragoneses <knocte@gmail.com>