Dap: better check for the playing track before ejecting (bgo#658142)
Before ejecting the device, there is some logic to stop playback
if the track that is currently playing is inside the device that
the user wants to eject. This logic was not taking in account that
this track can still be inside the device even if it's not playing
from the current DapSource (i.e. it can be in FileSystemQueue source).

For these cases, we check simply the path of the track now to see
if it starts with the BaseDirectory of the source.

Signed-off-by: Andres G. Aragoneses <knocte@gmail.com>